+++ Tacta.EventStore/Repository/EventStoreRecord.cs
@@ -6,7 +6,7 @@ namespace Tacta.EventStore.Repository
     {
         public string AggregateId { get; set; }
         public int Version { get; set; }
-        public int Sequence { get; set; }
+        public long Sequence { get; set; }
         public Guid Id { get; set; }
         public DateTime CreatedAt { get; set; }
         public T Event { get; set; }
